app/requirements.txt
fastapi
sqlalchemy
bcrypt
PyJWT
psycopg2
python-multipart
uvicorn
pytest
httpx
pyotp
slowapi
jinja2
prometheus-client
loguru
starlette